`QueryManyEnumeratedIter`
`QueryManyIter` skips entities that don't match the given query so if you call `enumerate()` on it the enumeration of the returned items doesn't match an enumeration of the list of entities.

This commit copies `QueryMany` from `bevy_ecs::iter::query`, changes the name to `QueryManyEnumerated` and adds a usize field named `index`. The `index` field is incremented inside the `fetch_next_aliased_unchecked` for loop so the enumeration of each item corresponds to its position in the entity list.