Improved API for gradients
Added `RectPosition` and `RectPositionAxis` structs to representat points relative to a Rectangle.

Added an `Ellipse` struct that represents axis-aligned rectangles.

Added a `resolve_geometry` method to `LinearGradient` that calculates the gradient's starting point and base length.

Improved the `resolve_geometry` method for `RadialGradient` to support the different enum variants and non-centered positions.